GS+UI: save samples and tracks. Empty samples can't be edited and tools can be used on them. When an empty uifile is associate with a file, every empty samples based on that uifile is filled. gs.save must be modified to re-save empty files/samples

+++ gs/samples/sampleDelete.js
@@ -1,7 +1,7 @@
 "use strict";
 
 gs.sampleDelete = function( sample ) {
-	if ( sample ) {
+	if ( sample && sample.wsample ) { // check wsample for empty sample
 		gs.sampleSelect( sample, false );
 		gs.samples.splice( gs.samples.indexOf( sample ), 1 );
 		sample.delete();
